pyproject.toml
[project]
name = "election-data"
version = "0.1.0"
description = "cleaned and normalized US election data"
authors = [
    {name = "Nick Crews", email = "nicholas.b.crews@gmail.com"},
]
dependencies = [
    "ibis-framework[duckdb]>=9.0.0.dev665",
    "ibis-widget",
    "ipykernel>=6.29.5",
    "pytest>=8.3.3",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[tool.uv.sources]
ibis-widget = { git = "https://github.com/NickCrews/ibis-widget.git", branch = "main" }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"